[Refactor] 기존의 MapExt에 있던 중복되는 코드들을 Factory패턴으로 정리
