+++ app/src/main/java/com/golddog/mask_location/ext/MapExt.kt
@@ -6,6 +6,7 @@ import androidx.core.content.ContextCompat
 import com.golddog.mask_location.R
 import com.golddog.mask_location.entity.HospitalClinic
 import com.golddog.mask_location.entity.StoreSales
+import com.golddog.mask_location.ext.status.*
 import com.golddog.mask_location.ui.dialog.InfoBottomSheet
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.NaverMap
@@ -27,31 +28,6 @@ fun setMarkerInvisible(markers: List<Marker>) {
 fun setMarkerVisibility(marker: Marker, visibility: Boolean, naverMap: NaverMap) {
     if (visibility) marker.map = naverMap
     else marker.map = null
-}
-
-
-fun setMarkerImage(overlayImage: OverlayImage, marker: Marker) {
-    marker.icon = overlayImage
-}
-
-fun setStoreMarkerTag(
-    storeSales: StoreSales,
-    status: String,
-    colorCode: Int
-): SpannableString {
-    val storeName = "${storeSales.name}\n"
-    val tagString =
-        "${storeSales.name}\n${storeSales.address}\n${status}\n" +
-                "입고시간 : ${storeSales.stockAt}\n갱신시간 : ${storeSales.createdAt}"
-    val storeNameStart = 0
-    val storeNameEnd = storeName.length
-    val statusStart = tagString.indexOf(status)
-    val statusEnd = statusStart + status.length
-
-    return SpannableString(tagString)
-        .bold(storeNameStart, storeNameEnd)
-        .sizeUp(storeNameStart, storeNameEnd)
-        .color(colorCode, statusStart, statusEnd)
 }
 
 fun setHospitalClinicMarkerTag(
@@ -81,40 +57,15 @@ fun setStoreMarker(
     marker.position = LatLng(storeSales.lat, storeSales.lng)
 
     if (status == "plenty") {
-        setMarkerImage(OverlayImage.fromResource(R.drawable.marker_plenty), marker)
-        infoString = setStoreMarkerTag(
-            storeSales,
-            context.getString(R.string.plenty_status),
-            ContextCompat.getColor(context, R.color.marker_plenty)
-        )
+        infoString = StatusFactory.getStatus(PlentyStatusFactory(), context, marker, storeSales)
     } else if (status == "some") {
-        setMarkerImage(OverlayImage.fromResource(R.drawable.marker_some), marker)
-        infoString = setStoreMarkerTag(
-            storeSales,
-            context.getString(R.string.some_status),
-            ContextCompat.getColor(context, R.color.marker_some)
-        )
+        infoString = StatusFactory.getStatus(SomeStatusFactory(), context, marker, storeSales)
     } else if (status == "few") {
-        setMarkerImage(OverlayImage.fromResource(R.drawable.marker_few), marker)
-        infoString = setStoreMarkerTag(
-            storeSales,
-            context.getString(R.string.few_status),
-            ContextCompat.getColor(context, R.color.marker_few)
-        )
+        infoString = StatusFactory.getStatus(FewStatusFactory(), context, marker, storeSales)
     } else if (status == "empty") {
-        setMarkerImage(OverlayImage.fromResource(R.drawable.marker_empty), marker)
-        infoString = setStoreMarkerTag(
-            storeSales,
-            context.getString(R.string.empty_status),
-            ContextCompat.getColor(context, R.color.marker_none)
-        )
+        infoString = StatusFactory.getStatus(EmptyStatusFactory(), context, marker, storeSales)
     } else if (status == "break") {
-        setMarkerImage(OverlayImage.fromResource(R.drawable.marker_break), marker)
-        infoString = setStoreMarkerTag(
-            storeSales,
-            context.getString(R.string.break_status),
-            ContextCompat.getColor(context, R.color.marker_none)
-        )
+        infoString = StatusFactory.getStatus(BreakStatusFactory(), context, marker, storeSales)
     } else {
         marker.map = null
     }
